[naga wgsl] Move code to generate WGSL for types into `common`.
Move the `TypeContext` trait from `naga::back::wgsl` into
`naga::common::wgsl`. Adjust imports and publicity markers as needed.

There should be no changes to behavior in this commit, only code
motion.
